% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_concepts.R
\name{automaticCorrelationCombineConcepts}
\alias{automaticCorrelationCombineConcepts}
\title{Function for automatically combining concepts by hierarchy mapping}
\usage{
automaticCorrelationCombineConcepts(
  data,
  abstractionLevel = -1,
  minCorrelation = 0.7,
  maxDaysInBetween = 1,
  heritageDriftAllowed = FALSE
)
}
\arguments{
\item{data}{CohortContrastObject}

\item{abstractionLevel}{abstraction level to use for mapping}

\item{minCorrelation}{minimum correlation to use for automatic concept combining}

\item{maxDaysInBetween}{minimum days between concepts to use for automatic concept combining}

\item{heritageDriftAllowed}{boolean for allowing heritage drift (combining concepts from differing heritages)}
}
\description{
Function for automatically combining concepts by hierarchy mapping
}
