% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{buildStudyMetadata}
\alias{buildStudyMetadata}
\title{Build study metadata for CSV/JSON sidecar files}
\usage{
buildStudyMetadata(object, studyName = NULL)
}
\arguments{
\item{object}{Object to save}

\item{studyName}{Optional study name override}
}
\description{
Build study metadata for CSV/JSON sidecar files
}
\keyword{internal}
