% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CohortContrast.R
\name{calculate_data_features}
\alias{calculate_data_features}
\title{This filters data based on prevalence difference ratio and return the nHighestPrevalenceDifference greates differences}
\usage{
calculate_data_features(data, nHighestPrevalenceDifference)
}
\arguments{
\item{data}{Data list object}

\item{nHighestPrevalenceDifference}{Number of features with highest prevalence difference ratio to keep}
}
\description{
This filters data based on prevalence difference ratio and return the nHighestPrevalenceDifference greates differences
}
\keyword{internal}
