% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{calculate_inverse_dates}
\alias{calculate_inverse_dates}
\title{Function for calculating inverse dates}
\usage{
calculate_inverse_dates(
  observation_period_start_date,
  observation_period_end_date,
  cohort_start_date,
  cohort_end_date
)
}
\arguments{
\item{observation_period_start_date}{Observation period start date}

\item{observation_period_end_date}{Observation period end date}

\item{cohort_start_date}{Cohort inclusion period start date}

\item{cohort_end_date}{Cohort inclusion end date}
}
\description{
Function for calculating inverse dates
}
\keyword{internal}
