% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohort_contrast_features.R
\name{calculateDataFeatures}
\alias{calculateDataFeatures}
\title{This filters data based on prevalence difference ratio and return the nHighestPrevalenceDifference greates differences}
\usage{
calculateDataFeatures(data, nHighestPrevalenceDifference)
}
\arguments{
\item{data}{Data list object}

\item{nHighestPrevalenceDifference}{Number of features with highest prevalence difference ratio to keep}
}
\description{
This filters data based on prevalence difference ratio and return the nHighestPrevalenceDifference greates differences
}
\keyword{internal}
