% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CohortContrastSafe.R
\name{calculatePlotData}
\alias{calculatePlotData}
\title{Function to calculate plot data for CCObject}
\usage{
calculatePlotData(data, abstractionLevel = -1)
}
\description{
Function to calculate plot data for CCObject
}
\keyword{internal}
