% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{checkForCorrectRelationDefinitions}
\alias{checkForCorrectRelationDefinitions}
\title{Function to check if target and control schemas and tables are defined sufficiently}
\usage{
checkForCorrectRelationDefinitions(
  targetTableName = NULL,
  controlTableName = NULL,
  targetTableSchemaName = NULL,
  controlTableSchemaName = NULL,
  cohortsTableSchemaName = NULL,
  cohortsTableName = NULL,
  targetCohortId = NULL,
  controlCohortId = NULL,
  pathToCohortsCSVFile = NULL
)
}
\arguments{
\item{targetTableName}{Name of the table where target cohort is defined}

\item{controlTableName}{Name of the table where control cohort is defined}

\item{targetTableSchemaName}{Name of the schema where target cohort table is defined}

\item{controlTableSchemaName}{Name of the schema where control cohort table is defined}

\item{cohortsTableSchemaName}{Name of the schema where cohorts' table is defined}

\item{cohortsTableName}{Name of the table where cohorts are defined}

\item{targetCohortId}{The id for target cohort in cohorts' table}

\item{controlCohortId}{The id for control cohort in cohorts' table}

\item{pathToCohortsCSVFile}{The path to a CSV file that has data table for cohorts}
}
\description{
Function to check if target and control schemas and tables are defined sufficiently
}
\keyword{internal}
