% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CohortContrastSafe.R
\name{CohortContrastSynthetic}
\alias{CohortContrastSynthetic}
\title{Filter CohortContrast Analysis with safety measures, will guarantee small cell supression between smallCellSuppression and 2*smallCellSuppression}
\usage{
CohortContrastSynthetic(
  data,
  smallCellSuppression = 5,
  complName = NULL,
  activeConceptKNN = FALSE,
  pathToResults = getwd()
)
}
\arguments{
\item{data}{Returned list from CohortContrast call}

\item{smallCellSuppression}{The rule of five integer, results under the threshold will be removed}

\item{complName}{Name of the output file}

\item{activeConceptKNN}{Use active feature values for clustering patients}

\item{pathToResults}{Path to the results folder, can be project's working directory}
}
\description{
Filter CohortContrast Analysis with safety measures, will guarantee small cell supression between smallCellSuppression and 2*smallCellSuppression
}
\examples{
\dontrun{
control <- data.frame(
  cohort_definition_id = c(1, 1, 1, 1, 1),
  subject_id = c(5325, 3743, 2980, 1512, 2168),
  cohort_start_date = as.Date(c("1982-06-02", "1997-03-23",
   "2004-09-29", "2006-08-11", "1977-06-25")),
  cohort_end_date = as.Date(c("2019-03-17", "2018-10-07",
   "2018-04-01", "2017-11-29", "2018-11-22"))
)

target <- data.frame(
  cohort_definition_id = c(1, 1, 1, 1, 1),
  subject_id = c(4804, 4861, 1563, 2830, 1655),
  cohort_start_date = as.Date(c("1997-03-23", "1982-06-02",
   "1977-06-25", "2006-08-11", "2004-09-29")),
  cohort_end_date = as.Date(c("2018-10-29", "2019-05-23",
   "2019-04-20", "2019-01-14", "2019-05-24"))
)

control$cohort_definition_id = 100
target$cohort_definition_id = 500

cohort = rbind(control, target)

con <- DBI::dbConnect(duckdb::duckdb(), dbdir = CDMConnector::eunomia_dir("GiBleed"))
DBI::dbExecute(con, "CREATE SCHEMA IF NOT EXISTS example")
DBI::dbWriteTable(con,   DBI::SQL('"example"."cohort"'), cohort)

cdm <- CDMConnector::cdm_from_con(con, cdm_name = "eunomia",
 cdm_schema = "main", write_schema = "main")

 targetTable <- cohortFromCohortTable(cdm = cdm, db = con,
  tableName = "cohort", schemaName = 'example', cohortId = 500)
controlTable <- cohortFromCohortTable(cdm = cdm, db = con,
 tableName = "cohort", schemaName = 'example', cohortId = 100)


pathToResults = getwd()

data = CohortContrast(
  cdm,
  targetTable = targetTable,
  controlTable = controlTable,
  pathToResults,
  domainsIncluded = c(
    "Drug"
  ),
  prevalenceCutOff = 0.1,
  topK = FALSE,
  presenceFilter = 0.005,
  complementaryMappingTable = NULL,
  runZTests = FALSE,
  runLogitTests = FALSE,
  createOutputFiles = FALSE,
  numCores = 1
)

CohortContrastSynthetic(data)

DBI::dbDisconnect(con)
}

}
