% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohort_contrast_viewer.R
\name{CohortContrastViewer}
\alias{CohortContrastViewer}
\title{CohortContrast Viewer - Interactive Disease Cohort Visualization}
\description{
Functions to configure Python environment and launch the CohortContrast Viewer dashboard
}
