% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdm_connector_setup.R
\name{cohortFromCSV}
\alias{cohortFromCSV}
\title{Read cohort from CSV}
\usage{
cohortFromCSV(pathToCsv, cohortId = NULL)
}
\arguments{
\item{pathToCsv}{Path to the cohort data CSV file}

\item{cohortId}{The id for cohort in cohorts' table, if NULL whole table will be imported}
}
\value{
a tbl object for further CohortContrast usage
}
\description{
Read cohort from CSV
}
\examples{
if (requireNamespace("readr", quietly = TRUE)) {
  pathToCsv <- tempfile(fileext = ".csv")
  cohort <- data.frame(
    cohort_definition_id = c(1L, 2L),
    subject_id = c(101L, 202L),
    cohort_start_date = as.Date(c("2020-01-01", "2020-02-01")),
    cohort_end_date = as.Date(c("2020-01-10", "2020-02-10"))
  )
  readr::write_csv(cohort, pathToCsv)

  targetTable <- cohortFromCSV(pathToCsv = pathToCsv, cohortId = 2)
  targetTable
}
}
