% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{convertToAbsolutePath}
\alias{convertToAbsolutePath}
\title{Convert a relative path to an absolute path}
\usage{
convertToAbsolutePath(path)
}
\arguments{
\item{path}{A string representing the file or directory path, which may be relative or absolute.}
}
\value{
A string with the absolute path. If the file or directory does not exist,
        `normalizePath()` will throw an error due to `mustWork = TRUE`.
}
\description{
This internal function converts a given relative path into an absolute path
using the built-in `normalizePath()` function.
}
\details{
The function utilizes `normalizePath()` to handle both Windows and Unix-based
         systems. It converts the provided path, even if it is relative, into
         an absolute path. It ensures that paths are formatted with forward slashes
         on Windows systems by setting `winslash = "/"`.
}
\examples{
\dontrun{
  relative_path <- "relative/path/to/file.txt"
  absolute_path <- convertToAbsolutePath(relative_path)
  print(absolute_path)
}
}
\keyword{internal}
