% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CohortContrast.R
\name{createC2TInputFunction}
\alias{createC2TInputFunction}
\title{Function for creating dataset which can be used as input for Cohort2Trajectory package}
\usage{
createC2TInputFunction(
  data,
  targetCohortId,
  createC2TInput,
  complementaryMappingTable = FALSE
)
}
\arguments{
\item{data}{Data list object}

\item{targetCohortId}{Target cohort id}

\item{createC2TInput}{Boolean for creating Cohort2Trajectory input dataframe}

\item{complementaryMappingTable}{Mappingtable for mapping concept_ids if present}
}
\description{
Function for creating dataset which can be used as input for Cohort2Trajectory package
}
\keyword{internal}
