% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdm_connector_setup.R
\name{createCohortContrastCdm}
\alias{createCohortContrastCdm}
\title{Insert cohort tables to CDM instance, preparing them from CohortContrast analysis}
\usage{
createCohortContrastCdm(cdm, targetTable = NULL, controlTable = NULL)
}
\arguments{
\item{cdm}{CDMConnector object}

\item{targetTable}{Table for target cohort (tbl)}

\item{controlTable}{Table for control cohort (tbl)}
}
\value{
object of dataframes and updated cdm object
}
\description{
Insert cohort tables to CDM instance, preparing them from CohortContrast analysis
}
\keyword{internal}
