% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CDMConnectorSetup.R
\name{createCohortContrastCohorts}
\alias{createCohortContrastCohorts}
\title{Generate cohort tables for CohortContrast analysis
Create relations of cohorts (target and control) to the database. If this function is run successfully, then CohortContrast package can be run.}
\usage{
createCohortContrastCohorts(
  cdm,
  db,
  targetTableName = NULL,
  controlTableName = NULL,
  targetTableSchemaName = NULL,
  controlTableSchemaName = NULL,
  cohortsTableSchemaName = NULL,
  cohortsTableName = NULL,
  targetCohortId = NULL,
  controlCohortId = NULL,
  pathToCohortsCSVFile = NULL,
  nudgeTarget = FALSE,
  nudgeControl = FALSE,
  useInverseControls = FALSE,
  useTargetMatching = FALSE
)
}
\arguments{
\item{cdm}{CDMConnector object}

\item{db}{Database instance (DBI)}

\item{targetTableName}{Name of the table where target cohort is defined}

\item{controlTableName}{Name of the table where control cohort is defined}

\item{targetTableSchemaName}{Name of the schema where target cohort table is defined}

\item{controlTableSchemaName}{Name of the schema where control cohort table is defined}

\item{cohortsTableSchemaName}{Name of the schema where cohorts' table is defined}

\item{cohortsTableName}{Name of the table where cohorts are defined}

\item{targetCohortId}{The id for target cohort in cohorts' table}

\item{controlCohortId}{The id for control cohort in cohorts' table}

\item{pathToCohortsCSVFile}{The path to a CSV file that has data table for cohorts}

\item{nudgeTarget}{number of days you would like to nudge the target cohort start day}

\item{nudgeControl}{number of days you would like to nudge the control cohort start day}

\item{useInverseControls}{Boolean for using inverse controls (target cohort's observation period which is not included)}

\item{useTargetMatching}{Boolean for using patient matching for controls}
}
\value{
object of dataframes and updated cdm object
}
\description{
Generate cohort tables for CohortContrast analysis
Create relations of cohorts (target and control) to the database. If this function is run successfully, then CohortContrast package can be run.
}
\examples{
# Example usage of createCohortContrastCohorts
control <- data.frame(
  cohort_definition_id = c(1, 1, 1, 1, 1),
  subject_id = c(5325, 3743, 2980, 1512, 2168),
  cohort_start_date = as.Date(c("1982-06-02", "1997-03-23",
   "2004-09-29", "2006-08-11", "1977-06-25")),
  cohort_end_date = as.Date(c("2019-03-17", "2018-10-07",
   "2018-04-01", "2017-11-29", "2018-11-22"))
)

target <- data.frame(
  cohort_definition_id = c(1, 1, 1, 1, 1),
  subject_id = c(4804, 4861, 1563, 2830, 1655),
  cohort_start_date = as.Date(c("1997-03-23", "1982-06-02",
   "1977-06-25", "2006-08-11", "2004-09-29")),
  cohort_end_date = as.Date(c("2018-10-29", "2019-05-23",
   "2019-04-20", "2019-01-14", "2019-05-24"))
)

control$cohort_definition_id = 100
target$cohort_definition_id = 500

cohort = rbind(control, target)

con <- DBI::dbConnect(duckdb::duckdb(), dbdir = CDMConnector::eunomia_dir("GiBleed"))
DBI::dbExecute(con, "CREATE SCHEMA IF NOT EXISTS testthat")
DBI::dbWriteTable(con,   DBI::SQL('"testthat"."cohort"'), cohort)

cdm <- CDMConnector::cdm_from_con(con, cdm_name = "eunomia",
 cdm_schema = "main", write_schema = "main")

cdm <- createCohortContrastCohorts(
  cdm,
  con,
  targetTableName = NULL,
  controlTableName = NULL,
  targetTableSchemaName = NULL,
  controlTableSchemaName = NULL,
  cohortsTableSchemaName = 'testthat',
  cohortsTableName = 'cohort',
  targetCohortId = 500,
  controlCohortId = 100,
  nudgeTarget = FALSE,
  nudgeControl = FALSE,
  useInverseControls = FALSE,
  useTargetMatching = FALSE
)
}
