% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helpers.R
\name{createComplementaryMappingTable}
\alias{createComplementaryMappingTable}
\title{Function for creating the complementary mapping table which can be used in CohortContrast execution function}
\usage{
createComplementaryMappingTable(conceptIds, conceptNames)
}
\arguments{
\item{conceptIds}{Array of concept ids to use}

\item{conceptNames}{Matching array of concept names to use}
}
\description{
Function for creating the complementary mapping table which can be used in CohortContrast execution function
}
\keyword{internal}
