% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CohortContrast.R
\name{createDataFeatures}
\alias{createDataFeatures}
\title{Function for extracting n features}
\usage{
createDataFeatures(data, topK)
}
\arguments{
\item{data}{Data list object}

\item{topK}{numeric > if set, keeps this number of features in the analysis. Maximum number of features exported.}
}
\description{
Function for extracting n features
}
\keyword{internal}
