% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualisations.R
\name{createHeatmap}
\alias{createHeatmap}
\title{Function for creating the heatmap which will be shown in shiny}
\usage{
createHeatmap(
  data,
  cohortDefinitionId,
  prevalenceRatioThreshold,
  prevalenceThreshold,
  complementaryMappingTable = NULL
)
}
\arguments{
\item{data}{data object returned by CohortContrast function}

\item{cohortDefinitionId}{Target cohort id}

\item{prevalenceRatioThreshold}{numeric > if set, removes all of the concepts which are not present (in target) more than prevalenceCutOff times}

\item{prevalenceThreshold}{numeric > if set, removes all features represented less than the given percentage}

\item{complementaryMappingTable}{Mappingtable for mapping concept_ids if present}
}
\description{
Function for creating the heatmap which will be shown in shiny
}
\keyword{internal}
