% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{createMandatorySubDirs}
\alias{createMandatorySubDirs}
\title{Function which creates mandatory subdirectories and files to the pathToResults directory}
\usage{
createMandatorySubDirs(pathToResults)
}
\arguments{
\item{pathToResults}{Path to the package results}
}
\description{
Function which creates mandatory subdirectories and files to the pathToResults directory
}
\keyword{internal}
