% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helpers.R
\name{createPathToResults}
\alias{createPathToResults}
\title{Function to ensure that the path to results exists, creating mandatory subdirectories if necessary}
\usage{
createPathToResults(pathToResults)
}
\arguments{
\item{pathToResults}{The path where results will be stored}
}
\description{
Function to ensure that the path to results exists, creating mandatory subdirectories if necessary
}
\keyword{internal}
