% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualisations.R
\name{createTargetMatrixForHeatmap}
\alias{createTargetMatrixForHeatmap}
\title{Function for creating matrix for the heatmap which will be shown in shiny}
\usage{
createTargetMatrixForHeatmap(
  data,
  targetCohortId,
  prevalenceCutOff,
  presenceFilter,
  complementaryMappingTable = FALSE
)
}
\arguments{
\item{data}{data object returned by CohortContrast function}

\item{targetCohortId}{Target cohort id}

\item{prevalenceCutOff}{numeric > if set, removes all of the concepts which are not present (in target) more than prevalenceCutOff times}

\item{presenceFilter}{numeric > if set, removes all features represented less than the given percentage}

\item{complementaryMappingTable}{Mappingtable for mapping concept_ids if present}
}
\description{
Function for creating matrix for the heatmap which will be shown in shiny
}
\keyword{internal}
