% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohort_contrast_viewer.R
\name{.getPythonDir}
\alias{.getPythonDir}
\title{Get the package's Python directory path}
\usage{
.getPythonDir()
}
\value{
Character string with the path to the Python code
}
\description{
Get the package's Python directory path
}
\keyword{internal}
