% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohort_contrast_viewer.R
\name{.getRequirementsPath}
\alias{.getRequirementsPath}
\title{Get the path to the requirements.txt file}
\usage{
.getRequirementsPath()
}
\value{
Character string with the path to requirements.txt
}
\description{
Get the path to the requirements.txt file
}
\keyword{internal}
