% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{.toJSON}
\alias{.toJSON}
\title{Function which converts text formatted JSON to digestible JSON}
\usage{
.toJSON(input, pretty = FALSE)
}
\arguments{
\item{input}{Text formatted JSON which needs conversion to digestible JSON}
}
\description{
Function which converts text formatted JSON to digestible JSON
}
\keyword{internal}
