% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/databaseFunctions.R, R/helpers.R
\name{dropRelation}
\alias{dropRelation}
\title{Function for deleting temporary tables from user's db}
\usage{
dropRelation(connection, dbms = "postgresql", schema = "", relationName)

dropRelation(connection, dbms = "postgresql", schema = "", relationName)
}
\arguments{
\item{connection}{Connection to the database (package DatabaseConnector)}

\item{dbms}{Database dialect}

\item{schema}{Schema in which the targeted table resides}

\item{relationName}{Name of the targeted table which will be dropped}
}
\description{
Function for deleting temporary tables from user's db

Function for deleting temporary tables from user's db
}
\keyword{internal}
