% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_helpers.R
\name{filter_target}
\alias{filter_target}
\title{Function for filtering CohortContrast results for plotting}
\usage{
filter_target(
  target,
  prevalence_thereshold = 0.01,
  prevalence_ratio_threshold = 1,
  domain,
  removeUntreated
)
}
\arguments{
\item{target}{Formatted CohortContrast object}

\item{prevalence_thereshold}{Prevalence threshold}

\item{prevalence_ratio_threshold}{Prevalence ratio threshold}

\item{removeUntreated}{Boolean for removing untreated patients}
}
\description{
Function for filtering CohortContrast results for plotting
}
\keyword{internal}
