% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdm_connector_setup.R
\name{filterTemporalBiasConcepts}
\alias{filterTemporalBiasConcepts}
\title{Filter temporal bias concepts from CohortContrast data}
\usage{
filterTemporalBiasConcepts(data, temporalBiasConcepts)
}
\arguments{
\item{data}{CohortContrast result object}

\item{temporalBiasConcepts}{Data frame of temporal bias concepts to remove}
}
\description{
Filter temporal bias concepts from CohortContrast data
}
\keyword{internal}
