% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_helpers.R
\name{format_results}
\alias{format_results}
\title{Function for formatting CohortContrast results for plotting}
\usage{
format_results(
  object,
  pathToResults,
  studyName,
  autoScaleRate,
  applyInverseTarget,
  applyZTest,
  applyLogitTest
)
}
\arguments{
\item{object}{CohortContrast returned object}

\item{pathToResults}{Path to the results folder, can be project's working directory}

\item{studyName}{Customized name for the study}

\item{autoScaleRate}{Boolean for scaling the prevalences}

\item{applyInverseTarget}{Boolean for applying inverse target}

\item{applyZTest}{Boolean for applying Z tests}

\item{applyLogitTest}{Boolean for applying Logit tests}
}
\description{
Function for formatting CohortContrast results for plotting
}
\keyword{internal}
