% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overwriteCohortGenerator.R
\name{generateCohort}
\alias{generateCohort}
\title{Generates a cohort, overwritten from original package CohortGenerator}
\usage{
generateCohort(
  cohortId = NULL,
  cohortDefinitionSet,
  connection = NULL,
  connectionDetails = NULL,
  cdmDatabaseSchema,
  cdmVocabSchema,
  tempEmulationSchema,
  cohortDatabaseSchema,
  cohortTableNames,
  stopIfError = TRUE,
  incremental,
  recordKeepingFile
)
}
\description{
Generates a cohort, overwritten from original package CohortGenerator
}
\keyword{internal}
