% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overwriteCohortGenerator.R
\name{generateCohortSet}
\alias{generateCohortSet}
\title{Generate a set of cohorts, overwritten from original package CohortGenerator}
\usage{
generateCohortSet(
  connectionDetails = NULL,
  connection = NULL,
  cdmDatabaseSchema,
  tempEmulationSchema = getOption("sqlRenderTempEmulationSchema"),
  cohortDatabaseSchema = cdmDatabaseSchema,
  cdmVocabSchema = cdmVocabSchema,
  cohortTableNames = getCohortTableNames(),
  cohortDefinitionSet = NULL,
  stopOnError = TRUE,
  incremental = FALSE,
  incrementalFolder = NULL
)
}
\description{
Generate a set of cohorts, overwritten from original package CohortGenerator
}
\keyword{internal}
