% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/databaseFunctions.R
\name{generateSourceTables}
\alias{generateSourceTables}
\title{Generate source code analysis tables}
\usage{
generateSourceTables(
  data,
  domainsIncluded = c("Drug", "Condition", "Measurement", "Observation", "Procedure",
    "Visit", "Visit detail", "Death")
)
}
\arguments{
\item{data}{CohortContrast object}

\item{domainsIncluded}{list of CDM domains to include}
}
\value{
object of dataframes and updated cdm object
}
\description{
Generate source code analysis tables
}
\keyword{internal}
