% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/databaseFunctions.R
\name{generateTables}
\alias{generateTables}
\title{Generate analysis tables
Create relations of cohorts (target and control) to the database}
\usage{
generateTables(
  cdm,
  domainsIncluded = c("Drug", "Condition", "Measurement", "Observation", "Procedure",
    "Visit", "Visit detail", "Death")
)
}
\arguments{
\item{cdm}{CDMConnector object: connection to the database}

\item{domainsIncluded}{list of CDM domains to include}
}
\value{
object of dataframes and updated cdm object
}
\description{
Generate analysis tables
Create relations of cohorts (target and control) to the database
}
\keyword{internal}
