% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_study_names}
\alias{get_study_names}
\title{Function for loading all of the study names saved in ./tmp/datasets}
\usage{
get_study_names(pathToResults)
}
\arguments{
\item{pathToResults}{Path to the results folder, can be project's working directory}
}
\description{
Function for loading all of the study names saved in ./tmp/datasets
}
\keyword{internal}
