% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualisations.R
\name{getHeatmapPlot}
\alias{getHeatmapPlot}
\title{Generate a Heatmap Plot}
\usage{
getHeatmapPlot(ccObject)
}
\arguments{
\item{ccObject}{An object returned by the CohortContrast GUI, typically created
when the "Create Snapshot" button is pressed. It should contain
target_row_annotation, target_col_annotation, and target_matrix components.}
}
\value{
A plot object generated by `plot_heatmap`.
}
\description{
This function generates a heatmap using the CohortContrast GUI object.
It extracts necessary components from the ccObject and passes them to the
`plot_heatmap` function to generate the plot.
}
\examples{
\dontrun{
# Assuming `ccObject` is already created via CohortContrast GUI
getHeatmapPlot(ccObject)
}
}
