% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jsonImport.R
\name{getJSONData}
\alias{getJSONData}
\title{Get data from OMOP CDM specified in imported JSON files}
\usage{
getJSONData(
  connection,
  connectionDetails,
  jsons,
  names,
  cdmDataSchema,
  cdmTempSchema,
  studyName
)
}
\arguments{
\item{connection}{Connection to database}

\item{connectionDetails}{An object of class connectionDetails as created by the DatabaseConnector::createConnectionDetails function.}

\item{jsons}{JSONs from which cohorts will be created}

\item{names}{Names of the cohorts}

\item{cdmDataSchema}{Schema which contains the OHDSI Common Data Model.}

\item{cdmTempSchema}{Schema for temporary tables}

\item{studyName}{Customized name for the study}
}
\description{
This function outputs a dataframe with columns SUBJECT_ID, COHORT_DEFINITION_ID, COHORT_START_DATE, COHORT_END_DATE
}
\keyword{internal}
