% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CohortContrast.R
\name{handleMapping}
\alias{handleMapping}
\title{This function uses complementaryMappingTable to map concepts to custom names}
\usage{
handleMapping(data, complementaryMappingTable, abstractionLevel = -1)
}
\arguments{
\item{data}{Data list object}

\item{complementaryMappingTable}{Mappingtable for mapping concept_ids if present, columns CONCEPT_ID, CONCEPT_NAME, NEW_CONCEPT_ID, NEW_CONCEPT_NAME, ABSTRACTION_LEVEL,}

\item{abstractionLevel}{Level of abstraction, by default -1 (imported data level)}
}
\description{
This function uses complementaryMappingTable to map concepts to custom names
}
\keyword{internal}
