% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohort_contrast_tests.R
\name{handleTests}
\alias{handleTests}
\title{Function for handling tests}
\usage{
handleTests(
  data,
  targetCohortId,
  presenceFilter,
  runChi2YTests = TRUE,
  runLogitTests = TRUE,
  numCores = parallel::detectCores() - 1
)
}
\arguments{
\item{data}{Data list object}

\item{targetCohortId}{Target cohort id}

\item{presenceFilter}{numeric > if set, removes all features represented less than the given percentage}

\item{runChi2YTests}{Boolean for running the CHI2Y test (chi-squared test
for two proportions with Yates continuity correction).}

\item{runLogitTests}{boolean for logit-tests}
}
\description{
Function for handling tests
}
\keyword{internal}
