% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{idExists}
\alias{idExists}
\title{Function for controlling whether patient exists in a cohort}
\usage{
idExists(data, id)
}
\arguments{
\item{data}{A dataframe object with SUBJECT_ID values}

\item{id}{The subject ID}
}
\description{
Function for controlling whether patient exists in a cohort
}
\keyword{internal}
