% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{is.nan.data.frame}
\alias{is.nan.data.frame}
\title{Function for finding NaN values in a data.frame object}
\usage{
\method{is.nan}{data.frame}(data)
}
\arguments{
\item{data}{SQL data.frame object}
}
\description{
Function for finding NaN values in a data.frame object
}
\keyword{internal}
