% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jsonImport.R
\name{loadJsons}
\alias{loadJsons}
\title{Load settings of the study from trajectorySettings.csv according to the customized paramater studyName}
\usage{
loadJsons(studyName, pathToResults)
}
\arguments{
\item{studyName}{Customized name for the study}

\item{pathToResults}{Path to the working directory, must include 'inst' folder}
}
\description{
Load settings of the study from trajectorySettings.csv according to the customized paramater studyName
}
\keyword{internal}
