% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{loadRenderTranslateSql}
\alias{loadRenderTranslateSql}
\title{Load and translate SQL file or an explicit SQL query to desired dialect.}
\usage{
loadRenderTranslateSql(
  sql,
  dbms = "postgresql",
  warnOnMissingParameters = TRUE,
  output = FALSE,
  outputFile,
  ...
)
}
\arguments{
\item{sql}{SQL file name or SQL query}

\item{warnOnMissingParameters}{Should a warning be raised when parameters provided to this function do not appear in the parameterized SQL that is being rendered? By default, this is TRUE.}

\item{output}{Should there be a .sql file created of the result}

\item{outputFile}{Name of the output file}
}
\description{
Load and translate SQL file or an explicit SQL query to desired dialect.
}
\keyword{internal}
