% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CDMConnectorSetup.R
\name{nudgeCohorts}
\alias{nudgeCohorts}
\title{Function to nudge cohorts}
\usage{
nudgeCohorts(cdm, nudgeTarget, nudgeControl)
}
\arguments{
\item{cdm}{Connection to database}

\item{nudgeTarget}{number of days you would like to nudge the target cohort start day}

\item{nudgeControl}{number of days you would like to nudge the control cohort start day}
}
\description{
This function outputs a dataframe with columns SUBJECT_ID, COHORT_DEFINITION_ID, COHORT_START_DATE, COHORT_END_DATE
}
\keyword{internal}
