% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CohortContrast.R
\name{performKSAnalysis}
\alias{performKSAnalysis}
\title{This function performs kolmogorov-smirnov tests on the patient concept occurrence date data target vs uniform distribution}
\usage{
performKSAnalysis(
  data_features,
  targetCohortId,
  numCores = parallel::detectCores() - 1
)
}
\arguments{
\item{data_features}{Summarised data of features}

\item{targetCohortId}{Target cohort id}

\item{numCores}{Number of cores to allocate to parallel processing}
}
\description{
This function performs kolmogorov-smirnov tests on the patient concept occurrence date data target vs uniform distribution
}
\keyword{internal}
