% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CohortContrast.R
\name{performPrevalenceAnalysisLogistic}
\alias{performPrevalenceAnalysisLogistic}
\title{This function learns separate logistic regression models on the patient prevalence data target vs control}
\usage{
performPrevalenceAnalysisLogistic(
  data_patients,
  data_initial,
  targetCohortId,
  presenceFilter
)
}
\arguments{
\item{data_patients}{Prevalence data for patients}

\item{data_initial}{Imported cohort dataframe}

\item{targetCohortId}{Target cohort id}

\item{presenceFilter}{Presence filter 0-1, if 0.1 then feature has to be present for at least 10 percent of patients}
}
\description{
This function learns separate logistic regression models on the patient prevalence data target vs control
}
\keyword{internal}
