% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_helpers.R
\name{plot_heatmap}
\alias{plot_heatmap}
\title{Function for prevalence heatmap plot}
\usage{
plot_heatmap(filtered_target)
}
\arguments{
\item{filtered_target}{filtered target object}
}
\description{
Function for prevalence heatmap plot
}
\keyword{internal}
