% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_helpers.R
\name{plot_prevalence}
\alias{plot_prevalence}
\title{Function for prevalence plot}
\usage{
plot_prevalence(filtered_target)
}
\arguments{
\item{filtered_target}{filtered target object}
}
\description{
Function for prevalence plot
}
\keyword{internal}
