% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{printCustomMessage}
\alias{printCustomMessage}
\title{Function for logging prints}
\usage{
printCustomMessage(message)
}
\arguments{
\item{message}{Message to show}
}
\description{
Function for logging prints
}
\keyword{internal}
