% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/databaseFunctions.R
\name{processCSVData}
\alias{processCSVData}
\title{This function creates target and control tables into the CDMConnector instance from the CSV files}
\usage{
processCSVData(
  cdm,
  useInverseControls = FALSE,
  useTargetMatching = FALSE,
  insertedCSVTarget,
  insertedCSVControl,
  nudgeTarget,
  nudgeControl
)
}
\arguments{
\item{cdm}{Connection to database (CDMConnector object)}

\item{useInverseControls}{Boolean for using inverse controls (target cohort's observation period which is not included)}

\item{useTargetMatching}{Boolean for using patient matching for controls}

\item{insertedCSVTarget}{The dataframe for target cohort}

\item{insertedCSVControl}{The dataframe for control cohort}

\item{nudgeTarget}{number of days you would like to nudge the target cohort start day}

\item{nudgeControl}{number of days you would like to nudge the control cohort start day}
}
\description{
This function creates target and control tables into the CDMConnector instance from the CSV files
}
\keyword{internal}
