% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CohortContrast.R
\name{queryHeritageData}
\alias{queryHeritageData}
\title{This function outputs a list of domain specigic data which is filtered by the worfklow configuration}
\usage{
queryHeritageData(
  dataPatient,
  cdm,
  studyName,
  split_data,
  complementaryMappingTable = FALSE
)
}
\arguments{
\item{dataPatient}{Data of the selected patients}

\item{cdm}{Connection to database}

\item{split_data}{Data split by heritage}

\item{complementaryMappingTable}{Mappingtable for mapping concept_ids if present}
}
\description{
This function outputs a list of domain specigic data which is filtered by the worfklow configuration
}
\keyword{internal}
