% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CDMConnectorSetup.R
\name{queryMatchedControlPrevalences}
\alias{queryMatchedControlPrevalences}
\title{Query concept prevalences for matched control cohort}
\usage{
queryMatchedControlPrevalences(cdm, matchedControl, domainsIncluded)
}
\arguments{
\item{cdm}{CDMConnector object}

\item{matchedControl}{Matched control cohort table}

\item{domainsIncluded}{List of domains to include}
}
\description{
Query concept prevalences for matched control cohort
}
\keyword{internal}
