% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CohortContrast.R
\name{resolveConceptNameOverlap}
\alias{resolveConceptNameOverlap}
\title{This function resolves concept name/label overlaps}
\usage{
resolveConceptNameOverlap(data)
}
\arguments{
\item{data}{Data list object}
}
\description{
This function resolves concept name/label overlaps
}
\keyword{internal}
