% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C2TWrapper.R
\name{resolveMissingDates}
\alias{resolveMissingDates}
\title{Function for creating dataset which can be used as input for Cohort2Trajectory package}
\usage{
resolveMissingDates(data, afterStart = 1, beforeEnd = 1)
}
\arguments{
\item{data}{Cohort data with columns COHORT_DEFINITION_ID, SUBJECT_ID, COHORT_START_DATE, COHORT_END_DATE}

\item{afterStart}{integer to add to COHORT_START_DATE if COHORT_END_DATE is missing}

\item{beforeEnd}{integer to subtract from COHORT_END_DATE if COHORT_START_DATE is missing}
}
\description{
Function for creating dataset which can be used as input for Cohort2Trajectory package
}
\keyword{internal}
