% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runGUI.R
\name{runCohortContrastGUI}
\alias{runCohortContrastGUI}
\title{This function initiates the connection with database and starts Shiny application}
\usage{
runCohortContrastGUI(pathToResults = NULL)
}
\arguments{
\item{pathToResults}{Path to target directory where results will be saved}
}
\description{
This function initiates the connection with database and starts Shiny application
}
