% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C2TWrapper.R
\name{runTrajectoryViz}
\alias{runTrajectoryViz}
\title{This function start up TrajectoryViz package GUI with the C2T output}
\usage{
runTrajectoryViz(data)
}
\arguments{
\item{data}{Cohort2Trajectory output object}
}
\description{
This function start up TrajectoryViz package GUI with the C2T output
}
