% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helpers.R
\name{sanitize_single}
\alias{sanitize_single}
\title{Function to sanitize a single string}
\usage{
sanitize_single(input_string)
}
\arguments{
\item{input_string}{A  state label name}
}
\description{
Function to sanitize a single string
}
\keyword{internal}
