% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{sanitizeSingle}
\alias{sanitizeSingle}
\title{Function to sanitize a single string}
\usage{
sanitizeSingle(inputString)
}
\arguments{
\item{inputString}{A state label name}
}
\description{
Function to sanitize a single string
}
\keyword{internal}
