% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{save_object}
\alias{save_object}
\title{Function for saving summary tables to path}
\usage{
save_object(object, path)
}
\arguments{
\item{object}{Object to save}

\item{path}{Path to the file saved}
}
\description{
Function for saving summary tables to path
}
\keyword{internal}
