% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{save_object_metadata}
\alias{save_object_metadata}
\title{Save study metadata to CSV}
\usage{
save_object_metadata(object, path, studyName = NULL)
}
\arguments{
\item{object}{Object to save}

\item{path}{Path to CSV file}

\item{studyName}{Optional study name override}
}
\description{
Save study metadata to CSV
}
\keyword{internal}
