% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{save_object_metadata_json}
\alias{save_object_metadata_json}
\title{Save study metadata to JSON}
\usage{
save_object_metadata_json(object, path, studyName = NULL)
}
\arguments{
\item{object}{Object to save}

\item{path}{Path to JSON file}

\item{studyName}{Optional study name override}
}
\description{
Save study metadata to JSON
}
\keyword{internal}
