% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohort_contrast_output.R
\name{saveResult}
\alias{saveResult}
\title{Save workflow outputs as parquet files}
\usage{
saveResult(data, pathToResults)
}
\arguments{
\item{data}{Data list object}

\item{pathToResults}{Path to the results folder, can be project's working directory}
}
\description{
Save workflow outputs as parquet files
}
\keyword{internal}
