% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helpers.R
\name{scale_to_1_0}
\alias{scale_to_1_0}
\title{# Function to normalize and scale a vector}
\usage{
scale_to_1_0(x)
}
\arguments{
\item{x}{Array of numeric}
}
\description{
# Function to normalize and scale a vector
}
\keyword{internal}
