% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_helpers.R
\name{update_features}
\alias{update_features}
\title{Function for interactively updating plots}
\usage{
update_features(features, scaled_prev)
}
\arguments{
\item{features}{features}

\item{scaled_prev}{scaled prev}
}
\description{
Function for interactively updating plots
}
\keyword{internal}
