% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combineConcepts.R
\name{automaticCorrelationCombineConcepts}
\alias{automaticCorrelationCombineConcepts}
\title{Function for automatically combining concepts by hierarchy mapping}
\usage{
automaticCorrelationCombineConcepts(
  data,
  abstraction_level = -1,
  minCorrelation = 0.7,
  maxDaysInbetween = 1
)
}
\arguments{
\item{data}{CohortContrastObject}

\item{abstraction_level}{abstraction level to use for mapping}

\item{minCorrelation}{minimum correlation to use for automatic concept combining}

\item{maxDaysInbetween}{minimum days inbetween concepts to use for automatic concept combining}
}
\description{
Function for automatically combining concepts by hierarchy mapping
}
