% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combineConcepts.R
\name{automaticHierarchyCombineConcepts}
\alias{automaticHierarchyCombineConcepts}
\title{Function for automatically combining concepts by hierarchy mapping}
\usage{
automaticHierarchyCombineConcepts(
  data,
  abstraction_level = -1,
  minDepthAllowed = 0,
  allowOnlyMinors = FALSE
)
}
\arguments{
\item{data}{CohortContrastObject}

\item{abstraction_level}{abstraction level to use for mapping}

\item{minDepthAllowed}{integer for restricting the mapping, if a concept is part of a hierarchy tree then minDepthAllowed value will prune the tree from said depth value upwards}

\item{allowOnlyMinors}{allows only mappping if child has smaller prevalence than parent}
}
\description{
Function for automatically combining concepts by hierarchy mapping
}
