% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combineConcepts.R
\name{automaticHierarchyCombineConcepts}
\alias{automaticHierarchyCombineConcepts}
\title{Function for automatically combining concepts by hierarchy mapping}
\usage{
automaticHierarchyCombineConcepts(data, abstraction_level = -1)
}
\arguments{
\item{data}{CohortContrastObject}

\item{abstraction_level}{abstraction level to use for mapping}
}
\description{
Function for automatically combining concepts by hierarchy mapping
}
