% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C2TWrapper.R
\name{c2tCaller}
\alias{c2tCaller}
\title{This function acts as a Cohort2Trajectory wrpper}
\usage{
c2tCaller(data, pathToResults)
}
\arguments{
\item{data}{CohortContrast object}

\item{pathToResults}{Path where data should be saved to}
}
\description{
This function acts as a Cohort2Trajectory wrpper
}
