% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CohortContrast.R
\name{CohortContrast}
\alias{CohortContrast}
\title{Get data from OMOP CDM specified in imported JSON files}
\usage{
CohortContrast(
  cdm,
  pathToResults,
  studyName,
  domainsIncluded = c("Drug", "Condition", "Measurement", "Observation", "Procedure",
    "Visit", "Visit detail"),
  readFromCSV = TRUE,
  prevalenceCutOff = 10,
  topDogs = FALSE,
  presenceFilter = 0.005,
  nudgeControl = FALSE,
  nudgeTarget = FALSE,
  complementaryMappingTable = FALSE,
  createC2TInput = FALSE,
  useInverseControls = FALSE,
  useTargetMatching = FALSE,
  runZTests = TRUE,
  runLogitTests = TRUE
)
}
\arguments{
\item{cdm}{Connection to database}

\item{pathToResults}{Path to the results folder, can be project's working directory}

\item{studyName}{Customized name for the study}

\item{domainsIncluded}{list of CDM domains to include}

\item{readFromCSV}{boolean > if TRUE file from ./inst/CSV will be used as cohort data, otherwise JSONs will be used.}

\item{prevalenceCutOff}{numeric > if set, removes all of the concepts which are not present (in target) more than prevalenceCutOff times}

\item{topDogs}{numeric > if set, keeps this number of features in the analysis, top n prevalence difference}

\item{presenceFilter}{numeric > if set, removes all features represented less than the given percentage}

\item{nudgeControl}{number of days you would like to nudge the control cohort start day}

\item{nudgeTarget}{number of days you would like to nudge the target cohort start day}

\item{complementaryMappingTable}{Mappingtable for mapping concept_ids if present}

\item{createC2TInput}{Boolean for creating Cohort2Trajectory input dataframe}

\item{useInverseControls}{Boolean for using inverse controls (target cohort's observation period which is not included)}

\item{useTargetMatching}{Boolean for using patient matching for controls}

\item{runZTests}{boolean for Z-tests}

\item{runLogitTests}{boolean for logit-tests}
}
\description{
Get data from OMOP CDM specified in imported JSON files
}
\keyword{external}
