% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CohortContrast.R
\name{CohortContrast}
\alias{CohortContrast}
\title{Get data from OMOP CDM specified in imported JSON files}
\usage{
CohortContrast(
  connection,
  connectionDetails,
  cdmSchema,
  cdmVocabSchema,
  cdmTmpSchema,
  pathToResults,
  studyName,
  domainsIncluded = c("Drug", "Condition", "Measurement", "Observation", "Procedure",
    "Visit", "Visit detail"),
  generateTables = TRUE,
  readFromCSV = FALSE,
  prevalenceCutOff = 10,
  topDogs = FALSE,
  presenceFilter = 0.005,
  removeOutliers = TRUE,
  nudgeControl = FALSE,
  nudgeTarget = FALSE,
  complementaryMappingTable = FALSE,
  runPCA = FALSE,
  runPCAClusters = 4
)
}
\arguments{
\item{connection}{Connection to database}

\item{connectionDetails}{An object of class connectionDetails as created by the DatabaseConnector::createConnectionDetails function.}

\item{cdmSchema}{JSONs from which cohorts will be created}

\item{cdmVocabSchema}{Schema for temporary tables}

\item{studyName}{Customized name for the study}

\item{domainsIncluded}{list of CDM domains to include}

\item{generateTables}{boolean > if TRUE new tables will be generated to the database.}

\item{readFromCSV}{boolean > if TRUE file from ./inst/CSV will be used as cohort data, otherwise JSONs will be used.}

\item{prevalenceCutOff}{numeric > if set, removes all of the concepts which are not present (in target) more than prevalenceCutOff times}

\item{topDogs}{numeric > if set, keeps this number of features in the analysis, top n prevalence difference}

\item{presenceFilter}{numeric > if set, removes all features represented less than the given percentage}

\item{removeOutliers}{boolean > if true this removes outlier patients from the target cohort, might cause loss of features}

\item{nudgeControl}{number of days you would like to nudge the control cohort start day}

\item{nudgeTarget}{number of days you would like to nudge the target cohort start day}

\item{complementaryMappingTable}{Mappingtable for mapping concept_ids if present}

\item{cdmDataSchema}{Schema which contains the OHDSI Common Data Model.}
}
\description{
This function outputs a dataframe with columns SUBJECT_ID, COHORT_DEFINITION_ID, COHORT_START_DATE, COHORT_END_DATE
}
\keyword{internal}
