% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CDMConnectorSetup.R
\name{cohortFromCohortTable}
\alias{cohortFromCohortTable}
\title{Read cohort from database cohort table}
\usage{
cohortFromCohortTable(
  cdm,
  db,
  tableName = NULL,
  schemaName = NULL,
  cohortId = NULL
)
}
\arguments{
\item{cdm}{CDMConnector object}

\item{db}{Database instance (DBI)}

\item{tableName}{Name of the table where the cohort is defined}

\item{schemaName}{Name of the schema where the cohort table is defined}

\item{cohortId}{The id for cohort in cohorts' table, if NULL whole table will be imported}
}
\value{
a tbl object for further CohortContrast usage
}
\description{
Read cohort from database cohort table
}
\examples{
\dontrun{
targetTable <- cohortFromCohortTable(cdm = cdm, db = db,
 tableName = "cohort", schemaName = 'ohdsi_results', cohortId = 1389)
}
\dontrun{
targetTable <- cohortFromCohortTable(cdm = cdm, db = db,
 tableName = "asthma", schemaName = 'user_peter')
}
}
