% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CDMConnectorSetup.R
\name{cohortFromCSV}
\alias{cohortFromCSV}
\title{Read cohort from CSV}
\usage{
cohortFromCSV(pathToCsv, cohortId = NULL)
}
\arguments{
\item{pathToCsv}{Path to the cohort data CSV file}

\item{cohortId}{The id for cohort in cohorts' table, if NULL whole table will be imported}
}
\value{
a tbl object for further CohortContrast usage
}
\description{
Read cohort from CSV
}
\examples{
\dontrun{
pathToCsv = './cohorts.csv'
targetTable <- cohortFromCSV(pathToCsv = pathToCsv, cohortId = 2)
}
}
