% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CDMConnectorSetup.R
\name{cohortFromJSON}
\alias{cohortFromJSON}
\title{Read cohort from JSON}
\usage{
cohortFromJSON(pathToJSON, cdm, cohortId = NULL)
}
\arguments{
\item{pathToJSON}{Path to the cohort data JSON file}

\item{cdm}{Connection to the database (package CDMConnector)}

\item{cohortId}{The id for cohort in cohorts' table, if NULL whole table will be imported}
}
\value{
a tbl object for further CohortContrast usage
}
\description{
Read cohort from JSON
}
\examples{
\dontrun{
pathToJSON = './JSON/'
targetTable <- cohortFromJSON(pathToJSON = pathToJSON, cdm, cohortId = 2)
}
}
