% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CDMConnectorSetup.R
\name{createControlCohortInverse}
\alias{createControlCohortInverse}
\title{Function for creating automatic matches based on inverse control logic}
\usage{
createControlCohortInverse(cdm, targetTable)
}
\arguments{
\item{cdm}{Connection to the database (package CDMConnector)}

\item{targetTable}{A cohort tibble which contains subjects' cohort data}
}
\description{
Function for creating automatic matches based on inverse control logic
}
\examples{
\dontrun{createControlCohortInverse(cdm = cdm, targetTable = targetTable)}
}
