% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/databaseFunctions.R
\name{createControlCohortInverse}
\alias{createControlCohortInverse}
\title{Function for creating automatic matches based on inverse control logic}
\usage{
createControlCohortInverse(cdm)
}
\arguments{
\item{cdm}{Connection to the database (package CDMConnector)}
}
\description{
Function for creating automatic matches based on inverse control logic
}
\keyword{internal}
