% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/databaseFunctions.R
\name{createControlCohortMatching}
\alias{createControlCohortMatching}
\title{Function for creating automatic matches based on age and sex}
\usage{
createControlCohortMatching(cdm, ratio = 1)
}
\arguments{
\item{cdm}{Connection to the database (package CDMConnector)}

\item{ratio}{ratio for the number of matches generated}
}
\description{
Function for creating automatic matches based on age and sex
}
\keyword{internal}
