% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CohortContrast.R
\name{createDataFeatures}
\alias{createDataFeatures}
\title{Function for extracting n features}
\usage{
createDataFeatures(data, topDogs)
}
\arguments{
\item{data}{Data list object}

\item{topDogs}{numeric > if set, keeps this number of features in the analysis, top n prevalence difference}
}
\description{
Function for extracting n features
}
\keyword{internal}
