% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mappingTableCreation.R
\name{generateMappingTable}
\alias{generateMappingTable}
\title{Create a mapping table with predefined maximum abstraction level}
\usage{
generateMappingTable(
  abstractionLevel = 10,
  data = NULL,
  maxMinDataFrame = NULL
)
}
\arguments{
\item{abstractionLevel}{Maximum abstraction level allowed}

\item{data}{CohortContrastObject returned by CohortContrast function}

\item{maxMinDataFrame}{A dataframe with columns descendant_concept_id and max_levels_of_separation}
}
\description{
Create a mapping table with predefined maximum abstraction level
}
