% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mappingTableCreation.R
\name{generateMappingTable}
\alias{generateMappingTable}
\title{Create a mapping table with predefined mmaximum abstraction level}
\usage{
generateMappingTable(cdm, maxAbstraction = 10, data = NULL)
}
\arguments{
\item{cdm}{Connection to database}

\item{maxAbstraction}{Maximum abstraction level allowed}

\item{data}{CohortContrastObject returned by CohortContrast function}
}
\description{
Create a mapping table with predefined mmaximum abstraction level
}
