% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/databaseFunctions.R
\name{generateTables}
\alias{generateTables}
\title{Create relations of cohorts (target and control) to the database}
\usage{
generateTables(
  connection,
  dbms,
  cdmSchema = "ohdsi_cdm",
  cdmVocabSchema = "ohdsi_vocab",
  cdmTmpSchema = "ohdsi_tmp",
  pathToResults = getwd(),
  studyName = "SampleStudy",
  generateTables = TRUE,
  domainsIncluded = c("Drug", "Condition", "Measurement", "Observation", "Procedure"),
  domainLimit = 1000,
  readFromCSV = FALSE,
  nudgeTarget = FALSE,
  nudgeControl = FALSE
)
}
\arguments{
\item{connection}{DatabaseConnector object: connection to the database}

\item{dbms}{DatabaseConnector connection database dialect}

\item{cdmSchema}{OHDSI CDM data schema}

\item{cdmVocabSchema}{OHDSI CDM vocabulary schema}

\item{cdmTmpSchema}{OHDSI CDM temp schema, user has write permissions}

\item{pathToResults}{Path to the neede subfolders such as 'inst'}

\item{studyName}{Name of the study being run, will be used in file and relation names}

\item{generateTables}{Boolean for recreating table}

\item{domainsIncluded}{list of CDM domains to include}

\item{domainLimit}{nr of instances allowed from a domain}

\item{readFromCSV}{boolean > if TRUE file from ./inst/CSV will be used as cohort data, otherwise JSONs will be used.}

\item{nudgeTarget}{number of days you would like to nudge the target cohort start day}

\item{nudgeControl}{number of days you would like to nudge the control cohort start day}
}
\description{
Create relations of cohorts (target and control) to the database
}
\keyword{internal}
