% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CohortContrast.R
\name{handleFeatureSelection}
\alias{handleFeatureSelection}
\title{Function for handling feature selection}
\usage{
handleFeatureSelection(
  data,
  pathToResults,
  topK,
  prevalenceCutOff,
  targetCohortId,
  runZTests,
  runLogitTests,
  createOutputFiles = FALSE
)
}
\arguments{
\item{data}{Data list object}

\item{pathToResults}{Path to the results folder, can be project's working directory}

\item{topK}{numeric > if set, keeps this number of features in the analysis. Maximum number of features exported.}

\item{prevalenceCutOff}{numeric > if set, removes all of the concepts which are not present (in target) more than prevalenceCutOff times}

\item{targetCohortId}{Target cohort id}

\item{runZTests}{boolean for Z-tests}

\item{runLogitTests}{boolean for logit-tests}

\item{createOutputFiles}{Boolean for creating output files, the default value is TRUE}
}
\description{
Function for handling feature selection
}
\keyword{internal}
