% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CohortContrast.R
\name{handleFeatureSelection}
\alias{handleFeatureSelection}
\title{Function for handling feature selection}
\usage{
handleFeatureSelection(
  data,
  studyName,
  topDogs,
  prevalenceCutOff,
  targetCohortId,
  runZTests,
  runLogitTests
)
}
\arguments{
\item{data}{Data list object}

\item{topDogs}{numeric > if set, keeps this number of features in the analysis, top n prevalence difference}

\item{prevalenceCutOff}{numeric > if set, removes all of the concepts which are not present (in target) more than prevalenceCutOff times}

\item{targetCohortId}{Target cohort id}

\item{runZTests}{boolean for Z-tests}

\item{runLogitTests}{boolean for logit-tests}
}
\description{
Function for handling feature selection
}
\keyword{internal}
