% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CohortContrast.R
\name{handleTests}
\alias{handleTests}
\title{Function for handling tests}
\usage{
handleTests(
  data,
  targetCohortId,
  presenceFilter,
  runZTests = TRUE,
  runLogitTests = TRUE,
  runKSTests = TRUE,
  numCores = parallel::detectCores() - 1
)
}
\arguments{
\item{data}{Data list object}

\item{targetCohortId}{Target cohort id}

\item{presenceFilter}{numeric > if set, removes all features represented less than the given percentage}

\item{runZTests}{boolean for Z-tests}

\item{runLogitTests}{boolean for logit-tests}

\item{runKSTests}{boolean for KS-tests}
}
\description{
Function for handling tests
}
\keyword{internal}
