% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CDMConnectorSetup.R
\name{matchCohortsByAge}
\alias{matchCohortsByAge}
\title{Function for matching the control to target by age}
\usage{
matchCohortsByAge(cdm, sourceTable, tableToMatch, maxAllowedAgeDifference = 0)
}
\arguments{
\item{cdm}{Connection to the database (package CDMConnector)}

\item{sourceTable}{Table which is used as reference for matching}

\item{tableToMatch}{Table which is matched}

\item{maxAllowedAgeDifference}{Value for maximum allowed age difference to be mapped to}
}
\description{
Function for matching the control to target by age
}
