% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phaseMining.R
\name{nGramClusterSummarization}
\alias{nGramClusterSummarization}
\title{Function for summarizing ngrams from nGramDiscovery output}
\usage{
nGramClusterSummarization(result, top_n = 5)
}
\arguments{
\item{result}{output from nGramDiscovery}

\item{top_n}{integer for number of most frequent concept to show per cluster}
}
\description{
Function for summarizing ngrams from nGramDiscovery output
}
