% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phaseMining.R
\name{nGramDiscovery}
\alias{nGramDiscovery}
\title{Function for discovering ngrams from the extracted data from CohortContrast function}
\usage{
nGramDiscovery(data, collapse_size = 0)
}
\arguments{
\item{data}{CohortContrastObject returned by CohortContrast or GUI snapshot}

\item{collapse_size}{integer for days to use for collapse same concept ids}
}
\description{
Function for discovering ngrams from the extracted data from CohortContrast function
}
