% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/databaseFunctions.R
\name{resolveCohortTableOverlaps}
\alias{resolveCohortTableOverlaps}
\title{Resolve overlaps inside the cohort table}
\usage{
resolveCohortTableOverlaps(cohortTable, cdm)
}
\arguments{
\item{cohortTable}{A table with cohort table characteristics}

\item{cdm}{CDMConnector object: connection to the database}
}
\value{
A dataframe like cohort table with resolved overlaps
}
\description{
Resolve overlaps inside the cohort table
}
