% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C2TWrapper.R
\name{resolveMissingDates}
\alias{resolveMissingDates}
\title{Function for creating dataset which can be used as input for Cohort2Trajectory package}
\usage{
resolveMissingDates(data, afterStart = 1, beforeEnd = 1)
}
\arguments{
\item{data}{Cohort data with columns cohort_definition_id, subject_id, cohort_start_date, cohort_end_date}

\item{afterStart}{integer to add to cohort_start_date if cohort_end_date is missing}

\item{beforeEnd}{integer to subtract from cohort_end_date if cohort_start_date is missing}
}
\description{
Function for creating dataset which can be used as input for Cohort2Trajectory package
}
\keyword{internal}
