% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runGUI.R
\name{runGUI}
\alias{runGUI}
\title{This function initiates the connection with database and starts Shiny application}
\usage{
runGUI(pathToResults = NULL, studyName = "Cohort2Trajectory")
}
\arguments{
\item{pathToResults}{Path to target directory where results will be saved}

\item{cdmResultsSchema}{Schema which has the information about the cohorts created in Atlas}
}
\description{
This function initiates the connection with database and starts Shiny application
}
