% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runGUI.R
\name{runGUI}
\alias{runGUI}
\title{This function initiates the connection with database and starts Shiny application}
\usage{
runGUI(pathToResults = NULL, studyName = "CohortContrast")
}
\arguments{
\item{pathToResults}{Path to target directory where results will be saved}

\item{studyName}{The study to load first in the GUI}
}
\description{
This function initiates the connection with database and starts Shiny application
}
