% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{save_object_metadata}
\alias{save_object_metadata}
\title{Function for saving metadata table to path}
\usage{
save_object_metadata(object, path, studyName = NULL)
}
\arguments{
\item{object}{Object to save}

\item{path}{Path to the file saved}
}
\description{
Function for saving metadata table to path
}
\keyword{internal}
