% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CohortContrast.R
\name{saveResult}
\alias{saveResult}
\title{This function saves the resulting data object}
\usage{
saveResult(data, pathToResults)
}
\arguments{
\item{data}{Data list object}

\item{pathToResults}{Path to the results folder, can be project's working directory}
}
\description{
This function saves the resulting data object
}
\keyword{internal}
