% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohort_contrast_viewer.R
\name{stopCohortContrastViewer}
\alias{stopCohortContrastViewer}
\title{Stop CohortContrast Viewer Dashboard}
\usage{
stopCohortContrastViewer()
}
\value{
Invisibly returns TRUE if the server was stopped.
}
\description{
Stops a running CohortContrast Viewer dashboard instance.
}
\examples{
\dontrun{
stopCohortContrastViewer()
}

}
